+++ demo/package.json
@@ -14,6 +14,7 @@
     "@tanstack/router-plugin": "^1.91.1",
     "@types/react": "^19.0.1",
     "@types/react-dom": "^19.0.2",
+    "@types/react-syntax-highlighter": "^15.5.13",
     "@vitejs/plugin-react": "^4.3.4",
     "vite": "^6.0.3"
   },
@@ -21,6 +22,7 @@
     "@tanstack/react-router": "^1.91.2",
     "@tanstack/router-devtools": "^1.91.2",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-syntax-highlighter": "^15.6.1"
   }
 }
