+++ demo/src/routes/persistent.tsx
@@ -1,42 +1,17 @@
-import { createFileRoute, Link } from "@tanstack/react-router";
-import { createPersistentStore, useStore } from "@fransek/statekit";
+import { createFileRoute } from "@tanstack/react-router";
+import { CodeBlock } from "../components/CodeBlock";
+import { Persistent } from "../examples/Persistent";
+import code from "../examples/Persistent?raw";
 
 export const Route = createFileRoute("/persistent")({
   component: RouteComponent,
 });
 
-// Create the store
-const store = createPersistentStore(
-  // Provide a unique key to identify the store in storage
-  "count",
-  { count: 0 },
-  (set) => ({
-    increment: () => set((state) => ({ count: state.count + 1 })),
-    decrement: () => set((state) => ({ count: state.count - 1 })),
-  }),
-  {
-    // storage: "session",
-    // serializer: superjson,
-  },
-);
-
 function RouteComponent() {
-  // Use the store
-  const {
-    state: { count },
-    actions: { increment, decrement },
-  } = useStore(store);
-
   return (
     <>
-      <div className="grid grid-cols-3 text-center items-center">
-        <button onClick={decrement}>-</button>
-        <div aria-label="count">{count}</div>
-        <button onClick={increment}>+</button>
-      </div>
-      <Link target="_blank" href="/persistent">
-        Duplicate this tab
-      </Link>
+      <Persistent />
+      <CodeBlock>{code}</CodeBlock>
     </>
   );
 }
